refactor(aruba): decompose mixins.go into mixin_common, mixin_scoped, mixin_status
